refactor: 상금 계산 관련 테스트 코드 위치 변경
