+++ src/test/java/domain/blackjack/PlayerTest.java
@@ -1,11 +1,5 @@
 package domain.blackjack;
 
-import static domain.blackjack.TestHoldingCards.BLACK_JACK;
-import static domain.blackjack.TestHoldingCards.DEAD_CARDS;
-import static domain.blackjack.TestHoldingCards.ONLY_SEVEN_HEART;
-import static domain.blackjack.TestHoldingCards.ONLY_SIX_HEART;
-import static domain.blackjack.TestHoldingCards.TWO_SIX_CARDS;
-import static domain.blackjack.TestHoldingCards.WIN_CARDS_WITHOUT_ACE;
 import static domain.card.Card.ACE_HEART;
 import static domain.card.Card.EIGHT_HEART;
 import static domain.card.Card.JACK_HEART;
@@ -29,37 +23,6 @@ class PlayerTest {
     static Stream<Arguments> validatePlayerHasNextDrawChanceParameters() {
         return Stream.of(
                 Arguments.of(TWO_HEART, false), Arguments.of(ACE_HEART, true)
-        );
-    }
-
-    public static Stream<Arguments> calculateBettingMoneyWhenDealerBustParameters() {
-        return Stream.of(
-                Arguments.of(BLACK_JACK, 10000, 15000),
-                Arguments.of(ONLY_SIX_HEART, 10000, 10000),
-                Arguments.of(DEAD_CARDS, 10000, 10000),
-                Arguments.of(WIN_CARDS_WITHOUT_ACE, 10000, 10000)
-        );
-    }
-
-    public static Stream<Arguments> calculateBettingMoneyWhenDealerNotBustParameters() {
-        return Stream.of(
-                Arguments.of(BLACK_JACK, 10000, 15000),
-                Arguments.of(ONLY_SIX_HEART, 10000, -10000),
-                Arguments.of(DEAD_CARDS, 10000, -10000),
-                Arguments.of(WIN_CARDS_WITHOUT_ACE, 10000, 10000),
-                Arguments.of(TWO_SIX_CARDS, 10000, 10000),
-                Arguments.of(ONLY_SEVEN_HEART, 10000, 0)
-        );
-    }
-
-    public static Stream<Arguments> calculateBettingMoneyWhenDealerIsBLackJackParameters() {
-        return Stream.of(
-                Arguments.of(BLACK_JACK, 10000, 0),
-                Arguments.of(ONLY_SIX_HEART, 10000, -10000),
-                Arguments.of(DEAD_CARDS, 10000, -10000),
-                Arguments.of(WIN_CARDS_WITHOUT_ACE, 10000, -10000),
-                Arguments.of(TWO_SIX_CARDS, 10000, -10000),
-                Arguments.of(ONLY_SEVEN_HEART, 10000, -10000)
         );
     }
 
@@ -92,41 +55,5 @@ class PlayerTest {
         Assertions.assertThatThrownBy(() -> Player.from("플레이어", HoldingCards.of(), -1))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("배팅 금액은 음수일 수 없습니다.");
-    }
-
-    @ParameterizedTest
-    @MethodSource("calculateBettingMoneyWhenDealerBustParameters")
-    @DisplayName("딜러가 버스트일 떼 배팅 상금이 제대로 계산되는지 검증")
-    void calculateBettingMoneyWhenDealerBust(HoldingCards playerHoldingCards, int bettingMoney, int expected) {
-        Player player = Player.from("플레이어", playerHoldingCards, bettingMoney);
-        Dealer dealer = Dealer.of(TestHoldingCards.DEAD_CARDS);
-        GameResult gameResult = GameResultCalculator.calculate(player, dealer);
-
-        Assertions.assertThat(EarnMoneyCalculator.calculateEarnMoney(bettingMoney, gameResult))
-                .isEqualTo(expected);
-    }
-
-    @ParameterizedTest
-    @MethodSource("calculateBettingMoneyWhenDealerNotBustParameters")
-    @DisplayName("딜러가 버스트가 아닐 때 배팅 상금이 제대로 계산되는지 검증")
-    void calculateBettingMoneyWhenDealerNotBust(HoldingCards playerHoldingCards, int bettingMoney, int expected) {
-        Player player = Player.from("플레이어", playerHoldingCards, bettingMoney);
-        Dealer dealer = Dealer.of(ONLY_SEVEN_HEART);
-        GameResult gameResult = GameResultCalculator.calculate(player, dealer);
-
-        Assertions.assertThat(EarnMoneyCalculator.calculateEarnMoney(bettingMoney, gameResult))
-                .isEqualTo(expected);
-    }
-
-    @ParameterizedTest
-    @MethodSource("calculateBettingMoneyWhenDealerIsBLackJackParameters")
-    @DisplayName("딜러가 블랙잭일 때 배팅 상금이 제대로 계산되는지 검증")
-    void calculateBettingMoneyWhenDealerIsBLackJack(HoldingCards playerHoldingCards, int bettingMoney, int expected) {
-        Player player = Player.from("플레이어", playerHoldingCards, bettingMoney);
-        Dealer dealer = Dealer.of(BLACK_JACK);
-        GameResult gameResult = GameResultCalculator.calculate(player, dealer);
-
-        Assertions.assertThat(EarnMoneyCalculator.calculateEarnMoney(bettingMoney, gameResult))
-                .isEqualTo(expected);
     }
 }
